gl_state: remove uniform buffer binding management
The uniform buffer binding point is not a single one but a generic one together with an array. With multiple uniform buffer needed now, it makes less sense to manage them by a single binding point. Also, the function glBindBufferBase used for binding to the array also mess up the state of generic binding point, making it impossible to manage by gl_state. Removing this management and manually manage the binding state for now, and in the future when the management is needed, a proper management for binding poing array is needed